# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Tournament API",
    "title": "Tournament API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/leagues/{tournament_alias}/{season_alias}/{stage_alias}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/TournamentAlias"
        },
        {
          "$ref": "#/components/parameters/SeasonAlias"
        },
        {
          "$ref": "#/components/parameters/StageAlias"
        }
      ],
      "get": {
        "operationId": "getLeagues",
        "tags": [
          "leagues"
        ],
        "description": "get tournament season stage leagues",
        "responses": {
          "200": {
            "description": "league list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaguesResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "description": "Tournament or season or stage not found",
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/groups/{tournament_alias}/{season_alias}/{stage_alias}/{league_alias}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/TournamentAlias"
        },
        {
          "$ref": "#/components/parameters/SeasonAlias"
        },
        {
          "$ref": "#/components/parameters/StageAlias"
        },
        {
          "$ref": "#/components/parameters/LeagueAlias"
        }
      ],
      "get": {
        "operationId": "getGroups",
        "tags": [
          "groups"
        ],
        "description": "get tournament season stage leagues",
        "responses": {
          "200": {
            "description": "league list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "description": "Tournament or season or stage not found",
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "tournament_not_found"
          },
          "message": {
            "type": "string",
            "example": "tournament not found"
          }
        }
      },
...
      "League": {
        "type": "object",
        "properties": {
          "alias": {
            "type": "string",
            "example": "first_league"
          },
          "name": {
            "type": "string",
            "example": "Первая лига"
          },
          "strengthWeight": {
            "type": "integer",
            "example": 3
          }
        }
      },
      "Group": {
        "type": "object",
        "properties": {
          "alias": {
            "type": "string",
            "example": "male_2021-2022_second_group_round_first_league_group_A"
          },
          "shortName": {
            "type": "string",
            "example": "А"
          },
          "name": {
            "type": "string",
            "example": "Группа А"
          },
          "description": {
            "type": "string",
            "example": "Мужской турнир, сезон 2021-2022, Первая лига, Круг 2, Группа А"
          },
          "tournamentAlias": {
            "type": "string",
            "example": "male"
          },
          "seasonAlias": {
            "type": "string",
            "example": "2021-2022"
          },
          "stageAlias": {
            "type": "string",
            "example": "2021-2022_second_group_round"
          },
          "leagueAlias": {
            "type": "string",
            "example": "male_2021-2022_second_group_round_first_league"
          }
        }
      },
...
      "LeaguesResponse": {
        "type": "object",
        "properties": {
          "leagues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/League"
            }
          }
        }
      },
      "GroupsResponse": {
        "type": "object",
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Group"
            }
          }
        }
      },
...
    },
    "parameters": {
      "TournamentAlias": {
        "name": "tournament_alias",
        "in": "path",
        "description": "tournament alias",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "SeasonAlias": {
        "name": "season_alias",
        "in": "path",
        "description": "season alias",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "StageAlias": {
        "name": "stage_alias",
        "in": "path",
        "description": "stage alias",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "LeagueAlias": {
        "name": "league_alias",
        "in": "path",
        "description": "league alias",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Identity not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
...
    }
  }
}